Add portsnap to DeltaPorts
Use this rather than SVN to update ports as a basis for generation.
It generates INDEX-3 which we need and SVN does not have.

It is necessary to create /var/db/portsnap manually before running it.
There were some minor modifications required, the biggest change was
changing "mv snap/ files/" to "mv snap files".  The former doesn't work
on DragonFly