poudriere: Fix list jails, spaces in jail names, build logs
THe list jails functionality for hammer had a bug in it.
Apparently spaces in jail names are legal if you can get it through
the command line, but to avoid issues with assumptions just replace
automatically all spaces with underscores.

Also make build dump to logs and store those in poudriere data.