Make clashing git requests line up
The first bulk build started failing somewhere along the line.  It is
suspected that two slaves hit the repository at the same time and caused
it to get out of wack.

The 5 minute wait can be used later by cron job to halt all commits
until the global pushing is complete.  That is upcoming work.