Fix p5-GSSAPI (cascade to Kerberos5, ghostscript, and gd)
DragonFly doesn't have a base Kerberos implementation, so it needs to
be added to p5-GSSAPI.  However, Kerberos builds documentation by default,
Tex-style documentation, and that sucks in a whole bunch of depends.
That required fixing gd for ghostscript and ghostscript for tex-base.
Now it should all build (mostly) as designed.