Improve scripts, especially merge (Fix :U and :L)
Three scripts changes involve setting ownership to automaton.
On the merge script, the main change is to use sed to change :U} and :L} to
:tu} and :tu} during the merging.

There are so many patches that it's not worth doing individually.
Sed doesn't work until after the patching so it won't screw up any
patches that are already there.  In time, I'll remove those I guess.