Add rpath to pkg
Starting with v1.0.4, pkg started failing with errors like, "can't find
libpkg.so.0" and "can't find /var/run/ld-elf.so.hints".  It shouldn't be
having to reference an ldconfig product to know where its own library is.
Add an additional rpath of "/usr/lib:/usr/local/lib" to pkg binary so
it can find it's libraries.

That solved the errors we were seeing immediately.