emulators/qemu-devel: Work around makecontext behaviour on DragonFly.
On DragonFly, when using swapcontext to enter a new context initialized
with makecontext all signals are left blocked in the new context. Since
qemu then uses siglongjmp (without a previously saved signal mask) to
continue in the original context, all signals are still blocked.

The most simple workaround is to just save the signal mask in the sigsetjmp
before using swapcontext to enter the new ucontext.

Without this workaround the virtual-machine isn't getting interrupted
(e.g. by the timer-interrupt) when it is running in a busy loop
(like "while true; do echo a > /dev/null; done" in the shell),
because SIGUSR1 is blocked.