bare minimum support for lang/rust
Rust (for dragonfly as the only BSD) uses procfs to get the patch of
the executing program.  FreeBSD uses KERN_PROC_PATHNAME but for some
reason if DF piggybacks on that, it fails.

Allow procfs temporarily, but we need it gone (synth doesn't even
support it although it can, but I don't want to make that exception)