Fix a longstanding issue with named pipes in libuv
These are the same patches as those found in [1] for the shipped version
of libuv of nodejs. Recent port revisions of nodejs (www/node) start to
build with the libuv version from dports. This led to a regression of
this bug in nodejs (and npm), as we haven't fixed this in devel/libuv
(yet). This commit fixes this issue.

[1]: https://github.com/DragonFlyBSD/DeltaPorts/pull/452