math/clblas: Fix build.
Even if I don't approve boost usage in OpenCL stuff, well, it makes a good testcase,
so fix cripplings of gcc (other compilers would not handle it anyway).
CLOCK_REALTIME on dlfy is visible later in POSIX.

This installs extra file in case of gcc, so PLIST_FILES+= trick is ok?
or should i just patch so that pc file is not install even for gcc case?