x11-drivers/xf86-video-ati: Move back to be a newport. (#776)
There are several issues, one of them is the git patch I need to exclude
(reported by ftigeot that on his r600 card glamor causes issues) and
I do not want more surprises. So it is best to keep both ati and intel29
to be handled consistency. As for mesa update, so far i915 gpu hangs on
anything that involves OpenGL, in ftigeot case it was hard reset.