devel/libpci: Make basic functionality work without root privileges.
Wit this patch, libpci will fall back to opening /dev/pci read-only, if the
user can't open it read-write, and it will then use the PCIOCGETCONF and
PCIOCGETBAR ioctls to retrieve information, instead of doing "raw" PCIOCREAD
and PCIOCWRITE ioctls.

This is needed to make the libpci usage in the chromium browser work
correctly.