lang/spidermonkey52: Do not set SONAME
Remove the patch that will set the `SONAME` for the built library.  The
library to be installed is `libmozjs-52.so.1`, while the `SONAME` to be
set is `libmozjs-52.so.0`, which mismatches the actually library.  Thus
the programs (e.g., `sysutils/polkit`) that links to this library would
fail to find the correct library and refuse to start.

Thanks-to: peeter007