net/zerotier: Rework the patchset to make it actually working
The existing patchset only makes ZeroTier build but it doesn't work.
Well, one workaround is required to get it working: manually create the
tapX interface before starting ZeroTier.

The new patchset fixes the TAP code and make it now working the same way
as on FreeBSD.

This is a port of my previous PR to get DragonFly support in ZeroTier:
https://github.com/zerotier/ZeroTierOne/pull/1043