Updating port for net/rinetd
This patch updates the port for net/rinetd to match the 2024Q3 branch on the FreeBSD ports tree.
The patch that was failing to apply cleanly was a patch in the 'files' directory;
I believe I have included the diff to fix the problematic file in the correct spot,
but please let me know if I have not and I will make the appropriate change(s).